added revenue to beta form
